# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FuelAnchor API
  description: |
    Blockchain-based fuel voucher system for East Africa.
    
    ## Features
    - 🔐 JWT Authentication
    - 💰 FUEL Token Management on Stellar Blockchain
    - 🎫 Voucher Generation & Redemption
    - 📍 GPS-based Geofencing
    - 💳 Mobile Money Integration (M-Pesa, MTN MoMo)
    - 🌍 Cross-Border Payments (SEP-31)
    - 📊 Analytics & Reporting
    
    ## Base URL
    - **Testnet**: `https://api-testnet.fuelanchor.io/api/v1`
    - **Production**: `https://api.fuelanchor.io/api/v1`
  version: 1.0.0
  contact:
    name: FuelAnchor Team
    email: support@fuelanchor.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api-testnet.fuelanchor.io/api/v1
    description: Testnet
  - url: https://api.fuelanchor.io/api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        from_account:
          type: string
        to_account:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [payment, refund, voucher, credit_purchase]
        status:
          type: string
          enum: [pending, completed, failed, cancelled]
        stellar_tx_hash:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        created_at:
          type: string
          format: date-time
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: Get user transactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, cancelled]
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
                  page:
                    type: integer

    post:
      tags: [Transactions]
      summary: Create new transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_account, to_account, amount, type, latitude, longitude]
              properties:
                from_account:
                  type: string
                to_account:
                  type: string
                amount:
                  type: number
                type:
                  type: string
                  enum: [payment, refund, voucher, credit_purchase]
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
...